requirements.txt
grazie_api_gateway_client
py2cfg
networkx[default]
pydantic